+++ src/EditorFeatures/Core/EmbeddedLanguages/DateAndTime/DateAndTimeEmbeddedLanguageEditorFeatures.cs
@@ -6,7 +6,7 @@
 
 using Microsoft.CodeAnalysis.Editor;
 using Microsoft.CodeAnalysis.Editor.EmbeddedLanguages;
-using Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices;
+using Microsoft.CodeAnalysis.EmbeddedLanguages;
 
 namespace Microsoft.CodeAnalysis.Features.EmbeddedLanguages.DateAndTime
 {
