+++ src/EditorFeatures/Core/EmbeddedLanguages/Json/JsonEmbeddedLanguageEditorFeatures.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices;
+using Microsoft.CodeAnalysis.EmbeddedLanguages;
 using Microsoft.CodeAnalysis.Features.EmbeddedLanguages.Json.LanguageServices;
 
 namespace Microsoft.CodeAnalysis.Editor.EmbeddedLanguages.Json
