+++ src/Features/Core/Portable/EmbeddedLanguages/IEmbeddedLanguageFeatures.cs
@@ -4,7 +4,7 @@
 
 using Microsoft.CodeAnalysis.Completion.Providers;
 using Microsoft.CodeAnalysis.DocumentHighlighting;
-using Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices;
+using Microsoft.CodeAnalysis.EmbeddedLanguages;
 
 namespace Microsoft.CodeAnalysis.Features.EmbeddedLanguages
 {
