+++ src/Workspaces/Core/Portable/EmbeddedLanguages/EmbeddedLanguageCommentDetector.cs
@@ -8,7 +8,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Text.RegularExpressions;
 
-namespace Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices
+namespace Microsoft.CodeAnalysis.EmbeddedLanguages
 {
     /// <summary>
     /// Helps match patterns of the form: language=name,option1,option2,option3
