+++ src/Workspaces/Core/Portable/EmbeddedLanguages/EmbeddedLanguageInfo.cs
@@ -7,7 +7,7 @@ using Microsoft.CodeAnalysis.EmbeddedLanguages.VirtualChars;
 using Microsoft.CodeAnalysis.LanguageServices;
 using Microsoft.CodeAnalysis.Shared.Collections;
 
-namespace Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices
+namespace Microsoft.CodeAnalysis.EmbeddedLanguages
 {
     internal readonly struct EmbeddedLanguageInfo
     {
