+++ src/Workspaces/Core/Portable/EmbeddedLanguages/IEmbeddedLanguagesProvider.cs
@@ -7,7 +7,7 @@
 using System.Collections.Immutable;
 using Microsoft.CodeAnalysis.Host;
 
-namespace Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices
+namespace Microsoft.CodeAnalysis.EmbeddedLanguages
 {
     /// <summary>
     /// Service that returns all the embedded languages supported.  Each embedded language can expose
